fix(tests): satisfy functional package structure
